requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi[standard]
torch==2.3.1+cpu
torchvision==0.18.1+cpu
transformers
numpy
pillow
google-genai
dotenv
requests
uvicorn